Move type check assertion to Type namespace
