+++ src/Assertion/Hierarchy/IsAssignableToPrimitive.php
@@ -14,9 +14,9 @@ use ReflectionClass;
 use Vivarium\Assertion\Assertion;
 use Vivarium\Assertion\Exception\AssertionFailed;
 use Vivarium\Assertion\Helpers\TypeToString;
-use Vivarium\Assertion\String\IsBasicType;
+use Vivarium\Assertion\Type\IsBasicType;
 use Vivarium\Assertion\String\IsEmpty;
-use Vivarium\Assertion\String\IsPrimitive;
+use Vivarium\Assertion\Type\IsPrimitive;
 
 use function sprintf;
