+++ src/Assertion/Type/IsInterface.php
@@ -8,18 +8,19 @@
 
 declare(strict_types=1);
 
-namespace Vivarium\Assertion\String;
+namespace Vivarium\Assertion\Type;
 
 use Vivarium\Assertion\Assertion;
 use Vivarium\Assertion\Exception\AssertionFailed;
 use Vivarium\Assertion\Helpers\TypeToString;
+use Vivarium\Assertion\String\IsEmpty;
 use Vivarium\Assertion\Var\IsString;
 
-use function class_exists;
+use function interface_exists;
 use function sprintf;
 
 /** @template-implements Assertion<class-string> */
-final class IsClass implements Assertion
+final class IsInterface implements Assertion
 {
     /** @psalm-assert class-string $value */
     public function assert(mixed $value, string $message = ''): void
@@ -27,7 +28,7 @@ final class IsClass implements Assertion
         if (! $this($value)) {
             $message = sprintf(
                 ! (new IsEmpty())($message) ?
-                     $message : 'Expected string to be a class name. Got %s.',
+                     $message : 'Expected string to be interface name. Got %s.',
                 (new TypeToString())($value),
             );
 
@@ -43,6 +44,6 @@ final class IsClass implements Assertion
     {
         (new IsString())->assert($value);
 
-        return class_exists($value);
+        return interface_exists($value);
     }
 }
